Split privy and non privy code in two crates
