+++ nym-vpn-core/Cargo.toml
@@ -12,6 +12,7 @@ members = [
     "crates/nym-macos",
     "crates/nym-offline-monitor",
     "crates/nym-platform-metadata",
+    "crates/nym-privy",
     "crates/nym-utils",
     "crates/nym-routing",
     "crates/nym-setup",
@@ -191,10 +192,11 @@ nym-ipc = { path = "crates/nym-ipc" }
 nym-macos = { path = "crates/nym-macos" }
 nym-offline-monitor = { path = "crates/nym-offline-monitor" }
 nym-platform-metadata = { path = "crates/nym-platform-metadata" }
-nym-utils = { path = "crates/nym-utils" }
+nym-privy = { path = "crates/nym-privy" }
 nym-routing = { path = "crates/nym-routing" }
 nym-statistics = { path = "crates/nym-statistics" }
 nym-statistics-api-client = { path = "crates/nym-statistics-api-client" }
+nym-utils = { path = "crates/nym-utils" }
 nym-vpn-account-controller = { path = "crates/nym-vpn-account-controller" }
 nym-vpn-api-client = { path = "crates/nym-vpn-api-client" }
 nym-vpn-lib = { path = "crates/nym-vpn-lib" }
