+++ nym-vpn-core/crates/nym-privy/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "nym-utils"
+name = "nym-privy"
 version.workspace = true
 authors.workspace = true
 repository.workspace = true
@@ -13,10 +13,7 @@ workspace = true
 
 [dependencies]
 bip39.workspace = true
+hex.workspace = true
+sha2.workspace = true
 thiserror.workspace = true
 
-nym-privy.workspace = true
-nym-vpn-lib-types.workspace = true
-
-[dev-dependencies]
-bip39 = { workspace = true, features = ["rand"] }
