+++ nym-vpn-core/crates/nym-privy/src/error.rs
@@ -1,0 +1,11 @@
+// Copyright 2025 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: GPL-3.0-only
+
+#[derive(thiserror::Error, Debug, Clone, PartialEq)]
+pub enum PrivyError {
+    #[error(transparent)]
+    Hex(#[from] hex::FromHexError),
+
+    #[error(transparent)]
+    Bip39(#[from] bip39::Error),
+}
