+++ nym-vpn-core/crates/nym-utils/src/lib.rs
@@ -3,23 +3,18 @@
 
 use bip39::Mnemonic;
 use nym_vpn_lib_types::LoginSecret;
-use sha2::{Digest, Sha256};
 
 use crate::error::UtilsError;
 
 pub mod error;
 
 pub fn parse_secret(secret: &LoginSecret) -> Result<Mnemonic, UtilsError> {
-    let bytes_signature = match secret {
-        LoginSecret::Mnemonic(mnemonic) => return Ok(Mnemonic::parse(mnemonic)?),
-        LoginSecret::PrivyHexSignature(hex_signature) => hex::decode(hex_signature)?,
+    let mnemonic = match secret {
+        LoginSecret::Mnemonic(mnemonic) => Mnemonic::parse(mnemonic)?,
+        LoginSecret::PrivyHexSignature(signature) => {
+            nym_privy::hex_signature_to_mnemonic(signature)?
+        }
     };
-
-    let mut hasher = Sha256::new();
-    hasher.update(&bytes_signature);
-    let hashed_signature = hasher.finalize();
-
-    let mnemonic = Mnemonic::from_entropy(&hashed_signature)?;
 
     Ok(mnemonic)
 }
@@ -42,9 +37,6 @@ mod tests {
         let hex_signature = String::from(
             "a564a87ccbed5cb5be4929201e555f5b5e26cb01d300d621520d724e57c582c33fa374caf21fd0c5e3118d70d14894845a32acfee47da7f347a0b9a57cba07931c",
         );
-
         assert!(parse_secret(&LoginSecret::PrivyHexSignature(hex_signature)).is_ok());
-        assert!(parse_secret(&LoginSecret::PrivyHexSignature(String::from("invalidhex"))).is_err());
-        assert!(parse_secret(&LoginSecret::PrivyHexSignature(String::from("deadbeef"))).is_ok());
     }
 }
